[BE] MOVE: Kafka consume에 필요한 dto 폴더 생성 후 이동 #105
