+++ src/backend/chat-server/src/main/java/kickzo/stomp_chat/dto/PlaylistUpdateEvent.java
@@ -1,0 +1,15 @@
+package kickzo.stomp_chat.dto;
+
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class PlaylistUpdateEvent {
+	private Long roomId;
+	private List<PlaylistItem> playlistJson;
+}
