+++ src/backend/chat-server/src/main/java/kickzo/stomp_chat/dto/RoomEvent.java
@@ -1,4 +1,4 @@
-package kickzo.stomp_chat.event;
+package kickzo.stomp_chat.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
