--- src/backend/chat-server/src/main/java/kickzo/stomp_chat/event/PlaylistUpdateEvent.java
@@ -1,11 +1,0 @@
-package kickzo.stomp_chat.event;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class PlaylistUpdateEvent {
-	private Long roomId;
-	private String playlistJson;
-}
